+++ Common/StmDogovorPro.Domain.Base/ExcelBase/ElementBase/IHasNAS.cs
@@ -2,6 +2,6 @@
 {
     public interface IHasNAS
     {
-        decimal NAS { get; set; }
+        double NAS { get; set; }
     }
 }
