+++ Common/StmDogovorPro.Domain.Base/ExcelBase/ElementBase/IHasRRC.cs
@@ -2,6 +2,6 @@
 {
     public interface IHasRRC
     {
-        int RRC { get; set; }
+        double RRC { get; set; }
     }
 }
